septentrio: avoid compilation error
Even though all enum values fit into 13 bits, there's an error:
../../src/drivers/gnss/septentrio/sbf/messages.h:101:21: error: ‘septentrio::sbf::Header::id_number’ is too small to hold all values of ‘enum class septentrio::sbf::BlockID’ [-Werror]
This is with GCC 13.3.0 under Ubuntu 24.04

+++ src/drivers/gnss/septentrio/septentrio_fuzz_tests.cpp
@@ -45,7 +45,7 @@ TEST(SeptentrioTestDecoding, decodeMessage)
 	septentrio::sbf::message_t m{};
 	m.header.sync1 = '$';
 	m.header.sync2 = '@';
-	m.header.id_number = septentrio::sbf::BlockID::PVTGeodetic;
+	m.header.id_number = static_cast<uint16_t>(septentrio::sbf::BlockID::PVTGeodetic);
 	m.header.length = sizeof(m.header) + sizeof(septentrio::sbf::PVTGeodetic);
 	septentrio::sbf::PVTGeodetic pvt{};
 	pvt.latitude = 123;
